requirements.txt
xformers==0.0.18
matplotlib==3.7.1
mmcv_full==1.7.1
numpy==1.24.3
opencv_python==4.7.0.72
pafy==0.5.5
torch==2.0.0+cu117
torchvision==0.15.0+cu117
yt_dlp==2023.7.6
